+++ fw_victron/dbus_definitions.py
@@ -1,12 +1,21 @@
 #!/usr/bin/python3
 
+# List of supported devices' types
+DEV_IFACE_BMV = "com.victron.BMV"
+DEV_IFACE_BlueSolar_MPPT = "com.victron.BlueSolarMPPT"
+DEV_IFACE_BlueSolar_MPPT_VECan = "com.victron.BlueSolarMPPTVECan"
+DEV_IFACE_SmartSolar_MPPT = "com.victron.SmartSolarMPPT"
+DEV_IFACE_SmartSolar_MPPT_VECan = "com.victron.SmartSolarMPPTVECan"
+DEV_IFACE_Phoenix_Inverter = "com.victron.PhoenixInverter"
+DEV_IFACE_Phoenix_Smart_IP43_Charger = "com.victron.PhoenixSmartIP43Charger"
+DEV_IFACE_SmartShunt = "com.victron.SmartShunt"
+
 # List of all DBus objects' definitions
 # Defined using the specs at https://www.victronenergy.com/upload/documents/VE.Direct-Protocol-3.33.pdf
 
 DEV_DBUS_DESC_SmartSolar_MPPT = '''
 <node>
-  <interface name='{dbus_name}'>
-
+  <interface name='{dbus_iface}'>
     <property name="load_state" type="b" access="read">
       <annotation name="org.freedesktop.DBus.Property.EmitsChangedSignal" value="true"/>
     </property>
@@ -55,7 +64,6 @@ DEV_DBUS_DESC_SmartSolar_MPPT = '''
     <property name="product_id" type="s" access="read">
       <annotation name="org.freedesktop.DBus.Property.EmitsChangedSignal" value="true"/>
     </property>
-
     <property name="firmware_version" type="s" access="read">
       <annotation name="org.freedesktop.DBus.Property.EmitsChangedSignal" value="true"/>
     </property>
@@ -65,7 +73,6 @@ DEV_DBUS_DESC_SmartSolar_MPPT = '''
     <property name="panel_voltage" type="i" access="read">
       <annotation name="org.freedesktop.DBus.Property.EmitsChangedSignal" value="true"/>
     </property>
-
   </interface>
 </node>
 '''